continued refactoring files
